// web/proto/budget/v1/auth_connect.ts
// @generated by protoc-gen-connect-es v1.6.1 with parameter "target=ts,import_extension=none"
// @generated from file budget/v1/auth.proto (package budget.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { LoginRequest, LoginResponse, RefreshTokenRequest, RefreshTokenResponse, RegisterRequest, RegisterResponse, RequestPasswordResetRequest, RequestPasswordResetResponse, ResetPasswordRequest, ResetPasswordResponse } from "./auth_pb";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service budget.v1.AuthService
 */
export const AuthService = {
  typeName: "budget.v1.AuthService",
  methods: {
    /**
     * @generated from rpc budget.v1.AuthService.Register
     */
    register: {
      name: "Register",
      I: RegisterRequest,
      O: RegisterResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.AuthService.Login
     */
    login: {
      name: "Login",
      I: LoginRequest,
      O: LoginResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.AuthService.RefreshToken
     */
    refreshToken: {
      name: "RefreshToken",
      I: RefreshTokenRequest,
      O: RefreshTokenResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.AuthService.RequestPasswordReset
     */
    requestPasswordReset: {
      name: "RequestPasswordReset",
      I: RequestPasswordResetRequest,
      O: RequestPasswordResetResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.AuthService.ResetPassword
     */
    resetPassword: {
      name: "ResetPassword",
      I: ResetPasswordRequest,
      O: ResetPasswordResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;