// web/proto/budget/v1/category_connect.ts
// @generated by protoc-gen-connect-es v1.6.1 with parameter "target=ts,import_extension=none"
// @generated from file budget/v1/category.proto (package budget.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { CreateCategoryRequest, CreateCategoryResponse, DeleteCategoryRequest, DeleteCategoryResponse, GetCategoryRequest, GetCategoryResponse, ListCategoriesRequest, ListCategoriesResponse, UpdateCategoryRequest, UpdateCategoryResponse } from "./category_pb";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service budget.v1.CategoryService
 */
export const CategoryService = {
  typeName: "budget.v1.CategoryService",
  methods: {
    /**
     * @generated from rpc budget.v1.CategoryService.CreateCategory
     */
    createCategory: {
      name: "CreateCategory",
      I: CreateCategoryRequest,
      O: CreateCategoryResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.CategoryService.UpdateCategory
     */
    updateCategory: {
      name: "UpdateCategory",
      I: UpdateCategoryRequest,
      O: UpdateCategoryResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.CategoryService.DeleteCategory
     */
    deleteCategory: {
      name: "DeleteCategory",
      I: DeleteCategoryRequest,
      O: DeleteCategoryResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.CategoryService.GetCategory
     */
    getCategory: {
      name: "GetCategory",
      I: GetCategoryRequest,
      O: GetCategoryResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc budget.v1.CategoryService.ListCategories
     */
    listCategories: {
      name: "ListCategories",
      I: ListCategoriesRequest,
      O: ListCategoriesResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;